Split out ItemSpec into UnlockableItem.{h,cpp}
