+++ UI/TechTreeWnd.cpp
@@ -13,9 +13,10 @@
 #include "../util/OptionsDB.h"
 #include "../util/Directories.h"
 #include "../util/ScopedTimer.h"
-#include "../universe/Tech.h"
 #include "../universe/Effects.h"
 #include "../universe/Enums.h"
+#include "../universe/Tech.h"
+#include "../universe/UnlockableItem.h"
 #include "../universe/ValueRef.h"
 #include "../Empire/Empire.h"
 #include "TechTreeLayout.h"
