+++ parse/EnumParser.cpp
@@ -1,6 +1,7 @@
 #include "EnumParser.h"
 
 #include "../universe/Enums.h"
+#include "../universe/UnlockableItem.h"
 
 
 namespace qi = boost::spirit::qi;
