+++ parse/Parse.cpp
@@ -6,7 +6,7 @@
 #include "EnumParser.h"
 #include "ValueRefParser.h"
 
-#include "../universe/Tech.h"
+#include "../universe/UnlockableItem.h"
 #include "../util/Logger.h"
 #include "../util/Directories.h"
