+++ parse/TechsParser.cpp
@@ -9,6 +9,7 @@
 #include "../universe/Effect.h"
 #include "../universe/Species.h"
 #include "../universe/Tech.h"
+#include "../universe/UnlockableItem.h"
 #include "../universe/ValueRef.h"
 #include "../util/Directories.h"
