+++ server/ServerWrapper.cpp
@@ -17,6 +17,7 @@
 #include "../universe/Pathfinder.h"
 #include "../universe/Universe.h"
 #include "../universe/UniverseGenerator.h"
+#include "../universe/UnlockableItem.h"
 #include "../universe/Enums.h"
 #include "../universe/ValueRef.h"
