+++ universe/Enums.h
@@ -193,17 +193,6 @@ GG_ENUM(DiplomaticStatus,
     NUM_DIPLO_STATUSES
 )
 
-/** types of items that can be unlocked for empires */
-GG_ENUM(UnlockableItemType,
-    INVALID_UNLOCKABLE_ITEM_TYPE = -1,
-    UIT_BUILDING,               ///< a kind of Building
-    UIT_SHIP_PART,              ///< a kind of ship part (which are placed into hulls to make designs)
-    UIT_SHIP_HULL,              ///< a ship hull (into which parts are placed)
-    UIT_SHIP_DESIGN,            ///< a complete ship design
-    UIT_TECH,                   ///< a technology
-    NUM_UNLOCKABLE_ITEM_TYPES   ///< keep last, the number of types of unlockable item
-)
-
 /** Research status of techs, relating to whether they have been or can be researched */
 GG_ENUM(TechStatus,
     INVALID_TECH_STATUS = -1,
