+++ universe/Tech.cpp
@@ -2,6 +2,7 @@
 
 #include "Effect.h"
 #include "UniverseObject.h"
+#include "UnlockableItem.h"
 #include "ObjectMap.h"
 #include "../util/OptionsDB.h"
 #include "../util/Logger.h"
@@ -78,12 +79,6 @@ namespace {
 }
 
 namespace CheckSums {
-    void CheckSumCombine(unsigned int& sum, const ItemSpec& item) {
-        TraceLogger() << "CheckSumCombine(Slot): " << typeid(item).name();
-        CheckSumCombine(sum, item.type);
-        CheckSumCombine(sum, item.name);
-    }
-
     void CheckSumCombine(unsigned int& sum, const TechCategory& cat) {
         TraceLogger() << "CheckSumCombine(Slot): " << typeid(cat).name();
         CheckSumCombine(sum, cat.name);
@@ -309,35 +304,6 @@ unsigned int Tech::GetCheckSum() const {
     return retval;
 }
 
-///////////////////////////////////////////////////////////
-// ItemSpec                                              //
-///////////////////////////////////////////////////////////
-ItemSpec::ItemSpec() :
-    type(INVALID_UNLOCKABLE_ITEM_TYPE),
-    name()
-{}
-
-std::string ItemSpec::Dump(unsigned short ntabs) const {
-    std::string retval = "Item type = ";
-    switch (type) {
-    case UIT_BUILDING:      retval += "Building";   break;
-    case UIT_SHIP_PART:     retval += "ShipPart";   break;
-    case UIT_SHIP_HULL:     retval += "ShipHull";   break;
-    case UIT_SHIP_DESIGN:   retval += "ShipDesign"; break;
-    case UIT_TECH:          retval += "Tech"    ;   break;
-    default:                retval += "?"       ;   break;
-    }
-    retval += " name = \"" + name + "\"\n";
-    return retval;
-}
-
-bool operator==(const ItemSpec& lhs, const ItemSpec& rhs) {
-    return lhs.type == rhs.type &&
-    lhs.name == rhs.name;
-}
-
-bool operator!=(const ItemSpec& lhs, const ItemSpec& rhs)
-{ return !(lhs == rhs); }
 
 ///////////////////////////////////////////////////////////
 // TechManager                                           //
